refactor: extract CLI program construction for version testing
